+++ cmds/Moderation/kick.js
@@ -6,24 +6,26 @@ module.exports = {
     description: 'Kicks a user',
     execute(message, Member, args) {
         message.delete();
-        const member = message.mentions.members.first();
-        message.channel.send(`Kick ${member}?`).then((edittthis) => {
-            edittthis.react('✅')
-            edittthis.react('❎')
-            message.client.on('messageReactionAdd', async (reaction, user) => {
-                if (user.bot) {
-                    return
-                }
-                if (reaction.emoji.name === '✅') {
-                    member.kick();
-                }
-                if (reaction.emoji.name === '✅') {
-                    edittthis.edit(`Successfully kicked ${member}`);
-                }
-                if (reaction.emoji.name === '❎') {
-                    edittthis.delete();
-                }
+        if (member.hasPermission('KICK_MEMBERS')) {
+            const member = message.mentions.members.first();
+            message.channel.send(`Kick ${member}?`).then((edittthis) => {
+                edittthis.react('✅')
+                edittthis.react('❎')
+                message.client.on('messageReactionAdd', async (reaction, user) => {
+                    if (user.bot) {
+                        return
+                    }
+                    if (reaction.emoji.name === '✅') {
+                        member.kick();
+                    }
+                    if (reaction.emoji.name === '✅') {
+                        edittthis.edit(`Successfully kicked ${member}`);
+                    }
+                    if (reaction.emoji.name === '❎') {
+                        edittthis.delete();
+                    }
+                })
             })
-        })
-        },
+        }
+    },
 };
