+++ cmds/Moderation/purge.js
@@ -6,8 +6,10 @@ module.exports = {
   category: 'Moderation',
   description: 'Clears the chat',
   execute(message, Member, args) {
-    let deletenum = args[0]
-    message.channel.bulkDelete(deletenum)
-    message.channel.send(`Deleted ${deletenum} messages! :wastebasket:`)
+    if (member.hasPermission('MANAGE_MESSAGES')) {
+      let deletenum = args[0]
+      message.channel.bulkDelete(deletenum)
+      message.channel.send(`Deleted ${deletenum} messages! :wastebasket:`)
+    }
   },
 };
